whitelist-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
apscheduler==3.10.4
gspread==6.1.0
oauth2client==4.1.3
httpx==0.27.0
python-dotenv==1.0.1
